Move the IExecutorProperties class to the common module.
