+++ hippo4j-spring-boot/hippo4j-config-spring-boot-starter/src/main/java/cn/hippo4j/config/springboot/starter/config/ExecutorProperties.java
@@ -17,6 +17,8 @@
 
 package cn.hippo4j.config.springboot.starter.config;
 
+import cn.hippo4j.common.api.ExecutorNotifyProperties;
+import cn.hippo4j.common.api.IExecutorProperties;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -101,7 +103,7 @@ public class ExecutorProperties implements IExecutorProperties {
     /**
      * Notify
      */
-    private DynamicThreadPoolNotifyProperties notify;
+    private ExecutorNotifyProperties notify;
 
     /**
      * Nodes, application startup is not affect, change properties is effect
